pyproject.toml
[tool.poetry]
name = "perfcheck"
version = "0.1.0"
description = ""
authors = ["David Bradford <david.bradford@mongodb.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/..."

[tool.poetry.scripts]
perf-check = 'perfcheck.cli:main'

[tool.poetry.dependencies]
python = "^3.9"
Click = "^7.0"
"evergreen.py" = "^3.0.0"
pydantic = "^1.8.1"

[tool.poetry.dev-dependencies]
pytest = "^6"
pytest-black = "^0.3"
pytest-cov = "^2.8"
pytest-flake8 = "^1.0"
pytest-mypy = "^0.7"
pytest-pydocstyle = "^2.0"
pytest-sugar = "^0.9"
pytest-isort = "^1.2.0"
flake8-bugbear = "^20.1.4"
black = "^20.8b1"
pre-commit = "^2.6.0"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100

[tool.pytest.ini_options]
flake8-ignore = "W605 W503 W291 E203 E501 F821"
addopts = "--flake8 --black --mypy --isort --pydocstyle"
mongodb_fixture_dir = "tests/integration_tests/fixtures"
testpaths = [
    "src",
    "tests",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"